requirements.txt
SpeechRecognition-ForkedVersion==3.9.2